Fri, Mar 18, 2022, 10:50 PM +00:00

+++ extra/extra4.c
@@ -5,7 +5,7 @@ int main(void)
 {
     int n = get_int("How many scores: ");
 
-    float scores[n];
+    int scores[n];
 
     for (int i = 0; i < n; i++)
     {
